+++ src/main/java/poomasi/domain/member/dto/response/SignUpResponse.java
@@ -1,4 +1,4 @@
-package poomasi.domain.auth.signup.dto.response;
+package poomasi.domain.member.dto.response;
 
 public record SignUpResponse(String email, String message) {
 }
