+++ src/main/java/poomasi/domain/member/service/MemberService.java
@@ -1,13 +1,18 @@
 package poomasi.domain.member.service;
 
+import jdk.jfr.Description;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import poomasi.domain.member.dto.response.MemberResponse;
+import poomasi.domain.member.entity.LoginType;
 import poomasi.domain.member.entity.Member;
 import poomasi.domain.member.repository.MemberRepository;
+import poomasi.domain.member.dto.request.SignupRequest;
+import poomasi.domain.member.dto.response.SignUpResponse;
 import poomasi.global.error.BusinessException;
 
 import static poomasi.domain.member.entity.Role.ROLE_CUSTOMER;
@@ -20,6 +25,25 @@ import static poomasi.global.error.BusinessError.*;
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    @Description("카카오톡으로 먼저 회원가입이 되어 있는 경우, 계정 연동을 진행합니다. ")
+    @Transactional
+    public SignUpResponse signUp(SignupRequest signupRequest) {
+        String email = signupRequest.email();
+        String password = signupRequest.password();
+
+        memberRepository.findByEmail(email)
+                .ifPresent(member -> { throw new BusinessException(DUPLICATE_MEMBER_EMAIL); });
+
+        Member newMember = new Member(email,
+                passwordEncoder.encode(password),
+                LoginType.LOCAL,
+                ROLE_CUSTOMER);
+
+        memberRepository.save(newMember);
+        return new SignUpResponse(email, "회원 가입 성공");
+    }
 
     public MemberResponse getMemberById(Long memberId) {
         Member member = findMemberById(memberId);
